requirements.txt
django
djangorestframework
requests
psycopg2-binary>=2.9